refactor (NuGettier.Upm): move struct definition for 'MetaGen.Guid' to namespace 'NuGettier.Upm.MetaGen'
