+++ NuGettier.Upm/StringFactories/MetaFactory.cs
@@ -7,7 +7,7 @@ public static partial class MetaFactory
 {
     public static string GenerateFolderMeta(string seed, string dirname)
     {
-        var guid = new Guid(seed, dirname);
+        var guid = new MetaGen.Guid(seed, dirname);
         var metaTemplate = Handlebars.Compile(
             EmbeddedAssetHelper.GetEmbeddedResourceString("NuGettier.Upm.Templates.folder.meta")
         );
@@ -18,7 +18,7 @@ public static partial class MetaFactory
 
     public static string GenerateFileMeta(string seed, string filename)
     {
-        var guid = new Guid(seed, filename);
+        var guid = new MetaGen.Guid(seed, filename);
         var metaTemplate = Handlebars.Compile(
             Path.GetExtension(filename).EndsWith(".dll")
                 ? EmbeddedAssetHelper.GetEmbeddedResourceString("NuGettier.Upm.Templates.assembly.meta")
